requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asyncclick==8.1.8.0
asyncpg==0.30.0
blinker==1.9.0
certifi==2025.4.26
click==8.1.8
dictdiffer==0.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
Redis-Sentinel-Url==1.0.1
rich==14.0.0
rich-toolkit==0.14.5
rq==2.3.3
rq-dashboard==0.8.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomli_w==1.2.0
tomlkit==0.13.2
tortoise-orm==0.25.0
typer==0.15.3
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
